requirements.txt
 fastapi
 uvicorn
 sqlalchemy
 pydantic
 cryptography
 python-jose[cryptography]
python-multipart

